[refac] move webConfig and related class
