+++ src/main/java/org/hankki/hankkiserver/api/common/annotation/UserId.java
@@ -1,4 +1,4 @@
-package org.hankki.hankkiserver.auth;
+package org.hankki.hankkiserver.api.common.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
